refactor: move types to own file
