+++ apps/journey-service/src/api/journey/types.ts
@@ -1,0 +1,12 @@
+export interface Journey {
+    startTime: Date;
+    distance: number;
+    userId: number;
+    lastPosition?: Coordinates;
+    endTime?: Date;
+}
+
+export interface Coordinates {
+    lat: number;
+    long: number;
+}
